Web Application and Exploitation Distribution
The WAED is a lightweight linux distribution based on Debian 8. WAED is
pre-configured with various real-world vulnerable web applications in a
sandboxed environment using Docker containers. It includes pentesting
tools that aid in finding web application vulnerabilities. The
motivation behind this project is to provide an environment to learn
about web application’s vulnerabilities without the hassle of dealing
with complex configurations. Currently, there are around 18 vulnerable
applications installed in WAED.